Update message formatting for SolidusShipStation::Api::RequestError
Sentry uses `.message`[^1] to set the value on their error messages. The
method exists by default on `Exception` (from which `RequestError`
inherits).  By default it returns the message provided to the
initialiser of `RequestError`.  This PR introduces better formatting of
this message to allow it to be picked up better by Sentry.

[^1]: https://github.com/getsentry/sentry-ruby/blob/94c28fe9a7fd1c15a10eeeb0e1c4fcac099d4166/sentry-ruby/lib/sentry/interfaces/single_exception.rb#L7