Retry sync_shipments for non-rate limiting errors
Based on our experience with the ShipStation API at my current project
we sometimes see errors that are resolved on the next run of the
SyncShipmentsJob.  To avoid noise in our error reporting this commit
makes use of the retry_on mechanism to only report an error after it
occurs a certain number of times.